Make poster-map page static like all other pages
- Convert page.tsx from client component to async server component
- Fetch data at build time using getMapData() instead of runtime API call
- Create PosterMapClient.tsx for interactive D3 map and scroll button
- Remove /api/map endpoint (no longer needed)
- Remove src/lib/api.ts (was only used by the API endpoint)
- Consolidate MapOrg type to single source in lib/data/map.ts
